requirements.txt
pydantic>=2.0
langchain
langchain-openai
langchain-community
faiss-cpu
python-dotenv
requests
huggingface